add a test case for a ns.c stack depth issue
I attempted to introduce the following seemingly trivial improvement:

  diff --git lib/common/ns.c lib/common/ns.c
  index 7e5eccc67..abd6c0228 100644
  --- lib/common/ns.c
  +++ lib/common/ns.c
  @@ -36,8 +36,9 @@ static int x_val(edge_t * e, node_t * v, int dir);
   static void check_cycles(graph_t * g);
   #endif

  -#define LENGTH(e)		(ND_rank(aghead(e)) - ND_rank(agtail(e)))
  -#define SLACK(e)		(LENGTH(e) - ED_minlen(e))
  +static int length(edge_t *e) { return ND_rank(aghead(e)) - ND_rank(agtail(e)); }
  +
  +#define SLACK(e)		(length(e) - ED_minlen(e))
   #define SEQ(a,b,c)		((a) <= (b) && (b) <= (c))
   #define TREE_EDGE(e)	(ED_tree_index(e) >= 0)

and to my surprise the #2646 test case began segfaulting. The cause appears to
be that my compiler (GCC 11.4.0) decides not to inline the `length` function,
despite it being very simple. The stack overhead of function calls into this in
turn overflows the stack during network simplex.

Because this issue is quite subtle and would not have been caught by CI testing
(we do not run the slow tests like `test_2646` in CI), it seems sensible to
introduce a stronger safeguard around this.

Gitlab: #2646