Revert "fix: set '\N' default node label during parsing if not already set"
This reverts commit f2ee992b4b81ab99a83b2bf6eefceabdfce4c212. This change had
the unexpected effect of causing constructing a graph in memory vs from a file
to have differing outcomes. Zooming out, it does not seem to make sense that
node label default were (and still are) being treated differently to other
defaults. A different solution will need to be found for #2434 and #2436.

Gitlab: re-opens #2434, re-opens #2436, fixes #2705
Reported-by: Rob Hart <robhart@google.com>