ortho overlapSeg: remove 'return(…)' idiom
Using parens on the expression being returned is an idiom that has been used to
allow `return` to be defined as a function-like macro. Nowadays this is
considered a bad idea and not something intended to be done in the Graphviz code
base. It can also be a concern to C++ programmers, where this idiom has
unintended side effects (see 7c43f27b58e3f79792c80f8d73786134b0bb10fe for more
discussion).