gvpr: fix: compute 'DFLT_GVPRPATH' at runtime
When Graphviz is built and installed to a given directory, `DFLT_GVPRPATH` was
being hard coded to contain this installation directory. This prevented
relocatable installations. This is not the only barrier to relocatability, but
removing it gets us one step closer.

This change also fixes a bug wherein 518b19d976f3b355ce98ce327e55b71b7e80f797
added a slightly incorrect path for `DFLT_GVPRPATH` in the CMake build system.

The technique used in this code in this commit is based on how Smyrna locates
its supporting resources at runtime.

Gitlab: #2530