fix: use '/' as path separator on MinGW when finding self
e4e0f7434b5b8495d7e5d1047034689ab14daadd observed that C and Python appear to
have different ideas of the path separator under MinGW. But this was incorrect.
The actual discrepancy here was that we are using (non-MinGW) Windows APIs to
retrieve a path, which hence uses the Windows separator '\'.

Accounting for this discrepancy is not only a simplification for `gv_find_me`
callers but actually a fix for gvedit and smyrna that previously could not
derive a path to their resource directories from a path to themselves. The
reason for this is that they would receive a path like
“C:\my files\bin\gvedit.exe” but then try to look for '/' to find the parent
directory.