gvc: on MinGW, use native path separator when globbing
This looks a bit strange – using '/' as a path separator only to convert back to
'\' when running the actual glob – but this seems more consistent. That is, now
each platform is dealing exclusively with its native path separators except
within the `glob` implementation in this file. `glob` has one path for native
Windows and one path for MinGW.