neatogen: encapsulate half edges state in an object
This removes reliance on some global state, making progress towards thread
safety. Note that this code still relies on the global `sqrt_nsites`.

This change is also a step towards removing use of neatogen’s internal allocator
in lib/neatogen/mem.c that also relies on global state.

Reference counting of half edges is removed in favor of accumulating a
linked-list of allocated half edges and freeing them all at the end. Outstanding
half edges need to be tracked because not all are in the hash table and can be
freed through pointers there, so it seemed simpler to remove intermediate
counting and freeing of half edges and instead just uniformly free them all
finally.

`ELcleanup` is moved from `cleanup` in adjust.c into voronoi.c. This is to make
initialization and cleanup more symmetric, as it was observed that the allocated
half edges and hash table do not need to be maintained across calls to
`voronoi`.

Gitlab: #2558