replace use of 'convert' with 'PIL' in blank GIF test
This removes a (optional) dependency on ImageMagick when running the test suite,
allowing this test to be run on Windows as well.

This was validated by rewinding to d7a89f3d69a253053b65dca91926e519d5619591 and
confirming this different phrasing reproduces the bug, and then stepping
forwards to dac13346d7d84c411a537855e18e39f69373fc80 and confirming it validates
the bug is fixed there.

Gitlab: #2609