expr evaldyn: fix use-after-free during map updates
When deleting an item from a map or updating an item in a map, the logic here
was not accounting for the possibility that the deleted item (or the
being-overwritten item) can still be in use elsewhere. As far as I can tell,
this bug has been present ever since this functionality was first implemented in
a7a5734ef0e3bfcb1c2670baad973de2d3477433.

Note that this change does not introduce a memory leak because the items in maps
are allocated from vmalloc, an allocator that frees all live heap pointers on
destruction.

Gitlab: #2626, fixes #2707
Reported-by: Steve Roush