Revert "expr: use system allocator for 'Exassoc_t's stored in 'local' dictionary"
This reverts commit cdd4b4ea1a771b8e5980ea04b2858821eb511713. It turns out the
lifetime of these objects is _not_ well contained as was thought during the
original change. Pointers to these objects are sometimes still within live
expressions after they have been removed from the locals dictionary.

Though this was driven by #2707, this alone does not fix the issue. This merely
reverts it to triggering a heap use-after-free.

Gitlab: #2626, #2707
Reported-by: Steve Roush