util: add an improved generic list
lib/util/list.h already contains a generic list, implemented as a large macro
that expands to both a data type definition as well as functions for operating
on it. This has a number of weaknesses, in particular the generated symbols are
not easily grepped for or found by code browsing tools.¹ At the time this was
implemented, it was believed there were only two leading styles for generic data
structures like this:
  1. A single list type, using `void *` for the data.² Operating on this
     structure is type-unsafe, in that it is easy to accidentally append e.g. an
     `int` value to a list of `struct foo`s due to interfaces being
     `void *`-typed. The compiler cannot spot this kind of misuse.
  2. Multiple types and their associated functions, generated by a macro like
     the one in list.h. This style is type safe but loses the ability to easily
     search for symbols.
That is, your choices were greppable symbols xor type safety.

Recently your author has become aware of a third style³ that achieves both
greppable symbols and type safety. This commit attempts to recapitulate list.h
in this style, though diverging in scenarios where we can do something simpler
or need something more complex. The long term plan is to migrate all list.h uses
to list2.h and then rename list2.h to list.h.

Note that it is only necessary to typedef the parametrized list type if it needs
to be passed between functions, assigned to compatible types, etc. For more
contained usage (like that in test_list2.c), no typedef is necessary. The need
for these typedefs in any context will become unnecessary in C23.⁴

Some notable differences from list.h to call out:
  • The implementation macros underscore-suffix local variables. In list.h, this
    is unnecessary because the locals are within functions and thus have
    scoping. In list2.h, they are within macros and need to make some attempt to
    anticipate and dodge other symbols from the context in which the macros are
    expanded.
  • `LIST_APPEND` vs `*_try_append`: The overflow check is dropped because we
    now call `gv_recalloc` instead of `realloc`. `gv_recalloc` does its own
    overflow checking.
  • `LIST_APPEND` vs `*_try_append`: We assign the `gv_recalloc` result directly
    to `list->base` instead of to a local `base` because (1) we have no direct
    way of accessing the list element type (no `typeof`) in order to declare
    `base` and (2) the reallocation cannot fail, so we do not need to worry
    about preserving the original pointer.
  • Because the implementations are macros where list.h’s implementations were
    functions, they have the foot gun of multiple expansion of the parameters.
    Thus expressions with side effects cannot be used with list2.h where they
    could be with list.h. In practice, this is not a significant limitation and
    the implementations are upper cased to visually remind callers these are
    macros.
  • To use `free` as a destructor, the sentinel `LIST_DTOR_FREE` needs to be
    used in place of `free`. This is because the member being assigned here has
    a defined type, whereas in list.h it just needed to be something taking a
    parameter of a type that the element type would coerce to.
  • `LIST_GET` returns an lvalue whereas `*_get` returned an rvalue. Callers
    should still avoid using the `LIST_GET` result as an lvalue and should
    instead use `LIST_AT` if they really need an lvalue.
  • `LIST_CONTAINS` requires the needle to be something we can take the address
    of. This restriction can be dropped using some features that arrive in C23,
    but surveying the existing uses this does not seem a significant restriction
    for now.
  • `LIST_COPY` and `LIST_DETACH` have different calling conventions where they
    take an output parameter instead of returning a value.
  • Comparators for `LIST_SORT` take `void *` parameters. This seems like a step
    _backwards_ in terms of type safety, but when rolling this out it was
    discovered this actually lets us remove some -Wcast-qual squashing tricks.
    Overall this seems like an improvement for noise reduction.

As might be inferred from some of the items above, the lack of C23 features (in
particular `typeof`) necessitated the `scratch` member and some other syntactic
noise. When moving to C23, this new implementation could be significantly
simplified and improved.

¹ e.g. https://gitlab.com/graphviz/graphviz/-/issues/2629#note_2566560015
² `SparseMatrix_struct` in lib/sparse/SparseMatrix.h is an example of this
  technique.
³ https://danielchasehooper.com/posts/typechecked-generic-c-data-structures/
⁴ https://www.open-std.org/jtc1/sc22/wg14/www/docs/n3037.pdf