smyrna: replace 'attrs_t' with a list2.h list
This is part of a series of changes to replace the ungreppable list.h type(s)
with the improved list2.h type(s).

Note that this appears to be a minor step backwards in that the comparator now
must operate on `void *` instead of typed pointers. However the `name##_sort`
implementation in list.h technically causes Undefined Behavior by casting the
comparator to an incompatible type. There does not appear to be a
straightforward standards-compliant way to use a type-safe comparator in list2.h
so we revert to a `void *` comparator.