common init_graph: simplify 'ctx' initialization
The zeroing of the fields `N_nodes`, `N_edges`, `S_i` became unnecessary after
28db8c8a6a260383e85d2c8146ddc7fe66a20a60,
79ef9b9eed37a0dbf503df4d870c862c7a998149, and
387c0299ce183c1573a22aec3be9c150334f8ef7. Instead of being globals that need
resetting, these are now encapsulated in the `ctx` struct that is zeroed by the
caller. We could drop the zeroing altogether, but zeroing the entire struct on
entry makes this function easier to understand in isolation and more future
proof against mistakes during refactoring.