mv graphviz_version.h.cmake graphviz_version.h.in
This file is being treated as a CMake source by the `lint_cmake_format` CI job
even though it is not. It is a C source with `@…@` variables added. This
discrepancy leads the linting job to fail with internal cmakelang errors after
upcoming changes. To avoid this, let’s make it clearer the language in use
within this file is not CMake.