common make_label: fix empty HTML-like strings being consumed by HTML parser
An unintended side effect of 1d5bd47203339d9f179f2b68e22ff126373353cb was that
an empty regular string and an empty HTML-like string became distinguishable
from one another. A flow on effect from this was that creating a label from an
empty HTML-like string no longer de-duped the string into a previously created
empty regular string (there is always one of these in the string intern
dictionary) and then created a regular label. Instead the empty HTML-like string
would be passed to the HTML parser for interpretation. This parser in turn does
not expect to receive an empty string and errors out.

This unintended interaction resulted in the observable problem that
`… label=<> …` could no longer be processed. Graphs containing syntax like this
would be rejected.

Gitlab: fixes #2727
Reported-by: Tom Moog