API BREAK: cgraph: remove 'agcanonStr'
This function uses a static buffer internally and thus is thread-unsafe. Debian
Code Search finds no material uses of it in the wild¹ so its removal is unlikely
to cause significant downstream work.

In the long term, the `agstrcanon` function is not an ideal API either, given
users need to inspect the source of `agstrcanon_bytes` to find out how much
memory they need to allocate for the output buffer. A better API would be:

  int agstrcanon(const char *str, FILE *out);

that incrementally writes the canonicalized string.

Gitlab: #2558

¹ https://codesearch.debian.net/search?q=agcanonStr&literal=1