add a test case that Autotools and CMake agree on plugin version
This should prevent forgetting to update one of these while working on the
other.

The CMake build system using a plugin age of 0 is more than just a
simplification. It appears using a non-zero age would require us to compute our
own (platform-dependent) library version:
  • https://cmake.org/Bug/view.php?id=4383
  • https://gitlab.kitware.com/cmake/cmake/-/issues/17652
This seems to be the result of an incorrect design decision made within CMake,
overruling well motivated objections from Brad King. If we ever bump the plugin
age, we will likely need to confront this issue and work around it.

Gitlab: #2530, !4535