add a test case that Autotools and CMake agree on library SO versions
This should prevent forgetting to update one of these while working on the
other.

Note that this test case needed to translate between libtool version info¹ and
library semantic version² suffixes. These are commonly incorrectly assumed to
map 1-1.³ The translation used here is based on the libtool source:⁴

  …
  case $version_type in
  # correct linux to gnu/linux during the next big refactor
  darwin|freebsd-elf|linux|midnightbsd-elf|osf|qnx|windows|none)
    func_arith $number_major + $number_minor
    current=$func_arith_result
    age=$number_minor
    revision=$number_revision
  …

noting that this branch includes Cygwin and MinGW:⁵

  …
  case $host_os in
  …
  cygwin* | mingw* | windows* | pw32* | cegcc*)
    version_type=windows
  …

A useful tool for exploring version info mapping is:
  • https://verbump.de/
  • described at, https://blog.hartwork.org/posts/shared-library-version-bump/

The CMake build system using library ages of 0 is more than just a
simplification. It appears using a non-zero age would require us to compute our
own (platform-dependent) library version:
  • https://cmake.org/Bug/view.php?id=4383
  • https://gitlab.kitware.com/cmake/cmake/-/issues/17652
This seems to be the result of an incorrect design decision made within CMake,
overruling well motivated objections from Brad King. If we ever bump the library
ages, we will likely need to confront this issue and work around it.

Gitlab: #2530, !4535

¹ https://www.gnu.org/software/libtool/manual/html_node/Updating-version-info.html
² https://semver.org/
³ https://autotools.info/libtool/version.html
⁴ https://cgit.git.savannah.gnu.org/cgit/libtool.git/tree/build-aux/ltmain.in#n7241
⁵ https://cgit.git.savannah.gnu.org/cgit/libtool.git/tree/m4/libtool.m4#n2661