fix excess points when generating paths using dot layout
This bug was introduced in 51e64d85197f203e349d0f1d84a287154753e996. The new
generic list unintentionally changed the evaluation order of some expressions in
a way that had a functional impact:

  -  points_append(plist, points_get(plist, points_size(plist) - 1));
  -  points_append(plist, points_get(plist, points_size(plist) - 1));
  +  LIST_APPEND(plist, LIST_GET(plist, LIST_SIZE(plist) - 1));
  +  LIST_APPEND(plist, LIST_GET(plist, LIST_SIZE(plist) - 1));

Under the old list implementation, these calls were functions, but under the new
list implementation they expanded (in the first level) to:

  do {
    const size_t slot_ =
      gv_list_append_slot_(&(plist)->impl, sizeof((plist)->base[0]));
    (plist)->base[slot_] = (LIST_GET(plist, LIST_SIZE(plist) - 1));
  } while (0);
  do {
    const size_t slot_ =
      gv_list_append_slot_(&(plist)->impl, sizeof((plist)->base[0]));
    (plist)->base[slot_] = (LIST_GET(plist, LIST_SIZE(plist) - 1));
  } while (0);

To understand why this is problematic, note that `gv_list_append_slot_` is not a
pure function. It modifies the list’s idea of its own size. An implicit
assumption was that no external callers would ever see this inconsistent state,
but the expanded call to `LIST_SIZE` here made that not true. As a result, these
lines were appending the wrong points. The visible effect of this was erratic
edges.

The commit that introduced the new generic list,
20063f263a040e32d1b679f735797207032070a2, noted the possibility of issues like
this:

  Because the implementations are macros where list.h’s implementations were
  functions, they have the foot gun of multiple expansion of the parameters.
  Thus expressions with side effects cannot be used with list2.h where they
  could be with list.h. In practice, this is not a significant limitation and
  the implementations are upper cased to visually remind callers these are
  macros.

but focused on the risk of multiple expansion rather than what was occurring
here: unexpected expansion order with respect to sequence points.

While this commit fixes the problem in `LIST_APPEND`, this problem also affects:
  • `LIST_TRY_APPEND`
  • `LIST_PREPEND`
  • _not_ `LIST_SET`, though it should probably be adjusted out of paranoia
    anyway
  • `LIST_DETACH`, though it would likely to take malformed usage to actually
    hit the bug here
These will be addressed in upcoming commits.

Gitlab: fixes #2734
Reported-by: Tamir Bahar