util: fix expansion order of parameters to 'LIST_DETACH'
As discussed in the previous commits, the previous definition could lead to
surprises due to `item` being expanded _after_ usages that modify `list`. Though
in the case of `LIST_DETACH` it would most likely only be possible through
misuse. The changes for this macro specifically involved a counter-intuitive
rephrasing, but this seems to be the best local option. A further improvement
will be made in a future commit.

Gitlab: #2734