util: fix expansion order of parameters to 'LIST_TRY_APPEND'
As discussed in the previous commit, the previous definition could lead to
surprises due to `item` being expanded _after_ usages that modify `list`.

Note that the unit test added here actually passes both before and after this
change on my local machine. This appears to be because the old definition of
`LIST_TRY_APPEND` depended on implementation-defined behavior. Expressions like
`LIST_TRY_APPEND(&foo, … LIST_SIZE(&foo) …)` resulted in an expansion that
modified the list size on the left hand side of an assignment and read it on the
right hand side of the assignment. `=` does not constitute a sequence point
here, so results were platform dependent. The new definition is hopefully not
subject to this problem.

Gitlab: #2734