util: extract the core of 'LIST_TRY_APPEND' into a function
The original implementation of this macro in
20063f263a040e32d1b679f735797207032070a2 used a convoluted definition in order
to remain an expression, which constrained it from defining locals. Now that it
uses the `scratch` member to store the item to be appended, we can rephrase it
into something simpler and more readable.

This change also adds integer overflow checks that were challenging to implement
within the macro previously.

Gitlab: #2734