neatogen: reuse arena allocator for allocating sites
This is both a performance optimization (the allocator in lib/util/arena.c is
more efficient than that in lib/neatogen/mem.c) and also a further step away
from relying on global state. This change drops the reference counting scheme
employed and instead just lets all sites be freed at the end of `voronoi`.

Gitlab: #2558