bump plugin revision
Commits like cdfc64ff10d2248e3c030295dbd9503141b3d4a4 have, since the last
release, modified the source of neatogen that is statically linked into the
neato layout plugin. Following the libtool conventions,¹ this implies the
revision number should be incremented. This is an attempt to realign development
practices with a more rigorous treatment of ABI compatibility, as the plugin
revision has never been bumped in the past even when it should have been.

¹ https://www.gnu.org/software/libtool/manual/html_node/Updating-version-info.html,
  https://autotools.info/libtool/version.html,
  https://verbump.de/,
  https://blog.hartwork.org/posts/shared-library-version-bump/