bump revision of libutil-depending libraries
Since the last release, ae8cd3277db3600aceb90dda59744963c78778d6 and other
commits have modified libutil that is statically linked into other libraries
affected in this commit. Following the libtool conventions,¹ this implies the
revision number should be incremented. This is an attempt to realign development
practices with a more rigorous treatment of ABI compatibility.

¹ https://www.gnu.org/software/libtool/manual/html_node/Updating-version-info.html,
  https://autotools.info/libtool/version.html,
  https://verbump.de/,
  https://blog.hartwork.org/posts/shared-library-version-bump/