circogen reduce: fix read of uninitialized value 'listCopy.dst'
CodeChecker says:

  [HIGH] lib/circogen/blockpath.c:471:3: The left operand of '==' is a
    garbage value [core.UndefinedBinaryOperatorResult]
      LIST_COPY(&listCopy, &list);
      ^

What it is noticing is that `listCopy` is uninitialized on calling this macro
and the first thing the macro did was `listCopy.base == list.base`. This is
Undefined Behavior, though most compilers would not exploit it because this
statement is typically (intentionally) optimized out as a non-side-effecting
pure operation.

We could initialize `listCopy` but a more robust fix is to rephrase the macro
such that it does not matter whether `dst` is initialized or not.

Reported-by: CodeChecker 6.26.2