sfdpgen shorting_edge_label_nodes: remove manual management of 'irn', 'jcn'
CodeChecker says:

  [HIGH] lib/sfdpgen/spring_electrical.c:1057:10: Array access (from variable
    'irn') results in a null pointer dereference [core.NullDereference]
    irn[nz] = mask[i];
            ^

  [HIGH] lib/sfdpgen/spring_electrical.c:1064:14: Array access (from variable
    'irn') results in a null pointer dereference [core.NullDereference]
        irn[nz] = mask[i];
                ^

These both appear to be false positives. But determining this requires carefully
examining the two loops in this function. We can avoid this complexity for both
humans and automated tools (as well as avoiding the cost of repeated loops) by
using generic lists that expand on-demand.

Note that the two `LIST_SYNC` calls here are no-ops; both lists should already
be synchronized. These are included for robustness in case the list
implementation changes in future.

Reported-by: CodeChecker 6.26.2