gvpr: fix: remove confusion about whether `uopts` can be null
CodeChecker says:

  [HIGH] lib/gvpr/gvpr.c:943:8: Access to field 'flags' results in a dereference
    of a null pointer (loaded from variable 'uopts') [core.NullDereference]
    if ((uopts->flags & GV_USE_EXIT))
         ^

`gvpr_core` checks whether `uopts` is null in some places but then in other
places unconditionally dereferences it. This change refactors logic to assume
the passed in `uopts` may be null and gracefully handle this case. This issue
seems to have existed since the very first revision of Graphviz.

Reported-by: CodeChecker 6.26.2