CMake: introduce an option for controlling Quartz discovery
This allows users to more explicitly indicate whether they want a build with the
Quartz plugin or not.

Note that in contrast to the Autotools build system that defaults to disabling
Quartz discovery, this defaults to enabling it when the Quartz framework exists
(typically only on macOS). This seems more inline with downstream users’
expectations.

This should not affect downstream Homebrew¹ or Macports² as neither of them are
using the CMake build system.

¹ https://github.com/Homebrew/homebrew-core/blob/820ba8bf659895388018d075d531273e1f4c9e6f/Formula/g/graphviz.rb
² https://github.com/macports/macports-ports/blob/ed31382b32959979dc7b1b52e417a0153a8a1607/graphics/graphviz/Portfile

Gitlab: #2530, #2714, #2735, !4559
Gitlab: https://gitlab.com/graphviz/graphviz/-/issues/2726#note_2749237319