util vagxbprint: fix double 'va_end' of a 'va_list'
CodeChecker says:

  [MEDIUM] lib/util/agxbuf.h:242:3: va_end() is called on an uninitialized
    va_list [valist.Uninitialized]
    va_end(ap);
    ^

What it is observing is that there is a path through this function that results
in calling `va_end(ap)` despite all callers of this function assuming they are
responsible for `va_end`-ing the list.

This change bumps the revision of components that are statically linked against
libutil and have not had a revision increment since the last release.

Reported-by: CodeChecker 6.26.2