CI: parallelize test running on Windows
The duration of CI has increased to over an hour on the most expensive Windows
jobs. The hope is that parallelizing the test suite’s execution is one way to
bring this time down.

The main concerns identified with this change are:
  • Degradation in readability of test results.¹ It appears this can be worked
    around if it turns out to be a problem.
  • Pexpect’s failure to create PTYs when Pytest is running multithreaded.² This
    should not be an issue on Windows because we skip the Pexpect-based tests
    there.

Some timeouts need to be increased in this change due to the Gitlab VMs being
oversubscribed. It seems the Windows runners have 2 vCPUS, but they do not have
the same throughput as 2 real CPUs.

¹ https://gitlab.com/graphviz/graphviz/-/issues/2660#note_2499891488
² https://gitlab.com/graphviz/graphviz/-/merge_requests/4255#note_2399315989

Gitlab: #2660, #2733