debian: upgrade Debian compatibility level 8 → 9
When trying to add Ubuntu 25.10 CI jobs, it was discovered that Smyrna
dependencies are not packaged for Ubuntu 25.10 which will probably require
conditional syntax in debian/graphviz.install that in turn requires
compatibility level 13.¹ This change is a step towards upgrading to
compatibility level 13.

It looks like the `binary` target could be dropped in debian/rules:²

  - dh is aware of the usual dependencies between targets in debian/rules. So,
    "dh binary" will run any build, build-arch, build-indep, install, etc
    targets that exist in the rules file. There's no need to define an explicit
    binary target with explicit dependencies on the other targets.

However it is explicitly called by ci/build.sh so we need to leave it in place.

¹ https://manpages.debian.org/trixie/debhelper/debhelper.7.en.html#Substitutions_in_debhelper_config_files
² https://manpages.debian.org/trixie/debhelper/debhelper-compat-upgrade-checklist.7.en.html#v9