common: remove use of 'times' on non-Windows
The main advantage of `times` over `clock` is its ability to exclude time taken
by children:¹

  Times for terminated children (and their descendants) are added in at the
  moment `wait(2)` or `waitpid(2)` returns their process ID. In particular,
  times of grandchildren that the children did not wait for are never seen.

While this is a valid concern, this is not typically a factor in any of the
places these timers are used in the Graphviz code.

Problematically this code hard codes a frequency of 60Hz. That is, that there
are 60 jiffies² in a second. This has not been true on Linux for a long time,
with the introduction of dynamic frequency and then steps towards “tickless”³
mode. As a result, this code was returning incorrect values for timing
measurements.

The correct way to learn the frequency is:¹

  The number of clock ticks per second can be obtained using:

    sysconf(_SC_CLK_TCK);

but rather than repairing this code to understand this dynamism, this change
simplifies it to use the unified – and ISO standardized – `clock` interface on
all platforms.

This is part of a broader set of changes to minimize platform-specific
differences in Graphviz behavior.

¹ https://linux.die.net/man/2/times
² https://en.wikipedia.org/wiki/Jiffy_(time)#Computing
³ https://lwn.net/Articles/549580/