gen_version: decide stable vs development through Git history inspection
On release day, the release process currently entails:
  1. Rebase the previously posted release Merge Request (MR), force-push this,
     mark it for automatic merge
  2. Wait for this to merge
  3. Construct a return-to-development-series MR, post this, mark it for
     automatic merge
  4. Wait for this to merge
  5. Wait for the merge of (2) to pass on main
  6. Construct an MR to update the website

The dependencies of this look like:

  digraph {
    1 -> 2;
    2 -> 3;
    3 -> 4;
    2 -> 5;
    5 -> 6;
  }

Diagrammatically:

       ┌─► ③ ─► ④
  ① ─► ②
       └─► ⑤ ─► ⑥

For the 14.0.0 release, the long pole of this ((1) ─► … ─► (4)) took ~6.5h. This
is becoming unsustainable, requiring maintainers to devote this much time to
shepherding CI through this process.

The main factor in why manual attention is required at all is that there exists
a vulnerable state between (2) and (4). During this time, if another
not-return-to-development-series MR is merged, two Graphviz commits will believe
themselves to be the release commit and will race one another through CI to post
the release artifacts.

This change attempts to remove the vulnerable state. Any commits
(return-to-development-series or otherwise) merged after (2) will correctly
believe themselves not to be the release commit. Note that, as the changes to
DEVELOPERS.md imply, explicitly starting the new development series is no longer
required. That is, steps (3) and (4) in the above process are removed.

This change is not completely 1-1 with respect to how commits are now evaluated.
For example, considering the commit graph surrounding the last release:

                                                        old          new
  *   747f78b30cd1364d18426cbf35163f2110e3c571       14.0.3~dev…  14.0.3~dev…
  |\  Merge
  | |
  | * 2416d3e9769c8b1ba0298e53217d9217c163f0f1       14.0.3~dev…  14.0.3~dev…
  | | Start 14.0.3 development
  |/
  *   8fe0778817801d0d88406c28e22c4b170cdbd517       14.0.2       14.0.2
  |\  Merge
  | |
  | * 0748394ba498f432079b687aafa5465de1a28bd6       14.0.2       14.0.3~dev…
  | | Stable Release 14.0.2
  |/
  *   c0f35f7e3775dca2539b7d9f7e183831ed2cbbe4       14.0.2~dev…  14.0.2~dev…
  |\  Merge
  …

Commits like 0748394ba498f432079b687aafa5465de1a28bd6 are now determined to be
part of the next development series. While odd looking, there are no known
problems caused by this. The authoritative definition of each release will
remain the Git tags that will continue to point at the merge of the
changelog-updating commit (8fe0778817801d0d88406c28e22c4b170cdbd517).

This change adds significant complexity to the version script, something that
has been error prone in the past and difficult to test. There does not seem to
be a simpler alternative to this available at present, but this would be a good
opportunity for future simplification if ideas arise.

Gitlab: closes #2733