cdt dtstat: unconditionally reallocate 'Count'
The previous logic was attempting to only expand `Count` when it was not large
enough to fulfill the current request already. On contemporary platforms,
allocator calls – especially for infrequent debugging operations like this
statistics code – are not a significant performance cost. We can afford to
simplify this code to no longer reuse memory.

This is a slight change in externally visible semantics. Previously, a series of
2 calls to `dtstat` would leave the `dt_count` member of the `Dtstat_t *` from
the first call as either a dangling pointer or an alias to the `dt_count` of the
second call. Now, the `dt_count` from the first call is definitively a dangling
pointer. It seems extremely unlikely any users could be adversely affected by
this, as we are simply narrowing two possible outcomes to one. Though it is
undocumented in the man page, it seems it has always been an assumption that
calling `dtstat` is known to invalidate the results of any previous `dtstat`
calls.

It would be better to drop reuse of `Count` altogether and make a new
disconnected allocation each call (i.e. drop the `static` from `Count`). However
this changes the semantics of the API (users are now expected to free the
`dt_count` pointer they receive) in a subtle way that cannot be diagnosed by
contemporary compilers.

Though it was not the primary goal of this change, it squashes the following
CodeChecker 6.26.2 warnings:

  [HIGH] lib/cdt/dtstat.c:63:13: Array access (from variable 'Count') results in
    a null pointer dereference [core.NullDereference]
        Count[i] = 0;
                 ^

  [HIGH] lib/cdt/dtstat.c:78:14: Array access (from variable 'Count') results in
    a null pointer dereference [core.NullDereference]
          Count[i] = 0;
                   ^