sfdpgen stress_model: remove dead 'x' allocation branch
CodeChecker says:

  [HIGH] lib/sfdpgen/stress_model.c:27:8: Dereference of null pointer (loaded
    from variable 'x') [core.NullDereference]
      *x = gv_calloc(m * dim, sizeof(double));
         ^

It appears as if the guard on this branch should be `!*x` instead of `!x`, the
latter of which _guarantees_ that this will dereference a null pointer. Rather
than repair this, we can zoom out and see nothing calls into `stress_model` with
a null pointer and therefore this branch has never been taken/tested. So it is
simpler to remove this code entirely.

Reported-by: CodeChecker 6.26.2