neatogen stress_majorization_with_hierarchy: fix memory leak of 'coords'
CodeChecker says:

  [MEDIUM] lib/neatogen/constrained_majorization.c:429:21: Potential leak of
    memory pointed to by 'coords' [unix.Malloc]
          iterations = -1;
                        ^

What it is noticing is that this branch jumps to the `finish` label, skipping
deallocation of `coords`. This is hard to see because of the extremely large
size and large number of variables of the containing function.

In this change we can drop the `coords != NULL` guard for the loop that remains
in place because it is intended to protect the `free(coords[0])` call, not the
rest of this logic which is a no-op when `coords == NULL`.

Reported-by: CodeChecker 6.26.2