note that VLAs are not permitted
Variable-Length Arrays¹ (VLAs) generally have no safe usage because they are
almost always a user-controlled stack overflow. The exception to this is
pointers to VLAs that can be very useful in matrix code and would simplify some
of the lib/neatogen and/or lib/sparse code. However attempting to use them
reveals that they are unsupported by MSVC:

  lib\neatogen\constrained_majorization.c(55,16): error C2057: expected constant
    expression

They have a waffling history in ISO C:
  • <C99: unsupported
  • C99: support required
  • C11: support optional
  • C17: as for C11
  • C23: support required

¹ https://en.wikipedia.org/wiki/Variable-length_array