gen_version: fix: ignore uncommitted changes
Something unaccounted for in 2f1c189b9790107100b8c19d64766d8017f0e820 was that
the Cygwin CI jobs convert line endings prior to running the build flow. A side
effect of this conversion is that a number of lines in files in the repository –
crucially, including the lines of CHANGELOG.md – appear modified and
uncommitted. Trying to `git blame` these lines returns this note and a commit
hash of 000000…, thwarting what gen_version.py is attempting to probe.

This change asks the `git blame` command to examine the latest commit, rather
than the current checkout state.

Gitlab: fixes #2785