circogen find_pair_edges: use a generic list for 'neighbors_without'
This avoids manual memory management and having to think about an upper bound on
the size of this container. It also potentially saves memory, as the full extent
of this array was not always used.

The `(size_t)diff` casts introduced are known safe because they are within a
`diff > 0` conditional block.