circogen find_pair_edges: avoid dereferencing null 'tp'
CodeChecker says:

  [HIGH] lib/circogen/blockpath.c:184:3: Access to field 'data' results in a
    dereference of a null pointer (loaded from variable 'tp')
    [core.NullDereference]
      DEGREE(tp)++;
      ^

As far as I can tell, this is a true positive. In the `diff == no_pair_count`
branch of this logic, `has_pair_count` may be 0. That is, there may be no edges
with pairs and `neighbors_with[0] == NULL`. It is not clear to me what should
happen in this case. But this change at least avoids the null pointer
dereference that would occur.

Reported-by: CodeChecker 6.26.2