sfio sfprintf: inline and simplify a call to 'sfucvt'
The `sfucvt` code appears to have been written to deal with constraints that no
longer apply. This code is no longer run in a freestanding environment, nor does
it need to assume the absence of a stdio implementation from libc.

This change was validated by constructing this harness that tests many possible
inputs to this logic:

  #include <limits.h>
  #include <sfio/sfhdr.h>
  #include <stdio.h>
  #include <stdlib.h>
  #include <string.h>

  int main(void) {

    for (int i = 1;; ++i) {
      char buf[1024] = {0};
      char *sp = buf + 100;
      char *ssp;
      long long lv = i;

      long long nv1;
      sfucvt(lv, sp, nv1, ssp, long long, unsigned long long);

      char copy[sizeof(buf)] = {0};
      memcpy(copy, buf, sizeof(buf));

      memset(buf, 0, sizeof(buf));
      sp = buf + 100;
      lv = i;
      char scratch[41];
      const int nv2 = snprintf(scratch, sizeof(scratch), "%lld", lv);
      sp -= nv2;
      memcpy(sp, scratch, (size_t)nv2);

      if (memcmp(copy, buf, sizeof(buf)) != 0) {
        fprintf(stderr, "printed content mismatch for %d\n", i);
        return EXIT_FAILURE;
      }

      if (i == INT_MAX)
        break;
    }
    return 0;
  }

This is a step towards removing barriers to multithreading.

Gitlab: #2558