sfio sfprintf: inline and simplify another call to 'sfucvt'
The `sfucvt` code appears to have been written to deal with constraints that no
longer apply. This code is no longer run in a freestanding environment, nor does
it need to assume the absence of a stdio implementation from libc.

The harness in the previous commit message is assumed to be enough to validate
this change too.

This is a step towards removing barriers to multithreading.

Gitlab: #2558